### 2019-02-12 #282 @DomBennett
outsider -- submission
Submitting Author: Dom Bennett (@dombennett)
Repository:   https://github.com/AntonelliLab/outsider
Version submitted:   0.0.0
Version accepted: 0.1.0

---

```
Package: outsider
Title: Install and run programs, outside of R, inside of R
Version: 0.0.0
Authors@R: person("Dom", "Bennett", role = c("aut", "cre"),
    email = "dominic.john.bennett@gmail.com",
Maintainer: Dom Bennett <dominic.john.bennett@gmail.com>
Description: Install and run external command-line programs in R through use of
    'Docker' <https://www.docker.com/> and 'GitHub' <http://github.com/>.
SystemRequirements: docker (>=18.0.0)
Language: en-GB
    utils (>= 3.1),
    crayon,
    devtools (>= 1.1),
    jsonlite (>= 1.1),
    sys (>= 2.1),
    yaml (>= 2.0),
    callr (>= 3.0.0),
    withr (>= 2.0),
    cli,
    praise
    testthat (>= 2.0),
    knitr (>= 1.0)
```

_The aim of the package is to enable R users to install and run, within R, software, packages and programs external to R. This allows users to more readily create workflows and pipelines that are portable, reproducible and make use of a wide range of available methods outside of R._

_The target audience are data scientists, bioinformaticians and general scientists interested in running external programs within R._

_There are many available packages that allow the interfacing between software external to R (e.g. packages for interacting with python or java). There does not, however, seem to be any packages that make use of docker (www.docker.com) to import external software as this package does. The closest packages to this one appear to be things like [rocker](https://github.com/rocker-org/rocker). Rocker, however, uses docker to make R more portable, not external software accessible within R._
