### 2019-02-16 #282 @annakrystalli
---

Many thanks again @DomBennett for your submission. A really interesting package!

I've completed the editor checks and am happy to start looking for reviewers. In the meantime, there are a couple of monior issues the initial checks have thrown up.

***

##### Test failure on `oldrel`

There's a test failure that pops up in examples when running on `oldrel`. Forking your repo and adding:

```
language: r
r:
  - oldrel
  - release
 ```
to `travis.yml` to also test on the old release of R indicates there's an API change between 3.4 and 3.5 R that affects package functionality, as oldrel fails.

In particular:

**Checks**

Throw an error in one of the `module_details` examples:

Running examples in ‘outsider-Ex.R’ failed
The error most likely occurred in:

> base::assign(".ptime", proc.time(), pos = "CheckExEnv")
> ### Name: module_details
> ### Title: Look up details on module(s)
> ### Aliases: module_details
>
> ### ** Examples
>
> library(outsider)
> # return table of ALL available modules on GitHub
> # NOT RUN - takes too long
> # (available_modules <- module_search())
>
> # look-up whether specific module exists
> # NOT RUN
> # repo <- 'dombennett/om..goodbye.world'
> # (suppressWarnings(module_exists(repo = repo)))
> repo <- 'dombennett/om..hello.world'
> (module_exists(repo = repo))
dombennett/om..hello.world
                      TRUE
> # get details on specific module(s)
> (module_details(repo = repo))
Error in if (inherits(X[[j]], "data.frame") && ncol(xj) > 1L) X[[j]] <- as.matrix(X[[j]]) :
  missing value where TRUE/FALSE needed
Calls: module_details -> lapply -> FUN -> as.matrix.data.frame
```
**Tests**

Again the same function fails the tests

* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ...
  Running ‘test-all.R’ [10s/74s]
 ERROR
Running the tests in ‘tests/test-all.R’ failed.
  Container status 'This is a mock'
  ────────────────────────────────────────────────────────────────────────────────────────
  ── 1. Error: module_details() works (@test-search.R#94)  ───────────────────────────────
  missing value where TRUE/FALSE needed
  1: module_details(repo = c(repo, repo)) at testthat/test-search.R:94
  2: lapply(X = github_res, FUN = as.matrix)
  3: FUN(X[[i]], ...)
  4: as.matrix.data.frame(X[[i]], ...)

  ```

I've checked the [CI](https://ropensci.github.io/dev_guide/ci.html) section in our guide and admittedly we don't have any guidance on this currently, so we are discussing updating it as a result. In the meantime, can I ask that you:

- [ ] set up `travis.yml` to also test on `devel`, `release` and `oldrel` (I can make a PR if you like)?
- [ ] address the errors in `oldrel`? If you need any help debugging, just let us know.

##### Good practice

Firstly, `goodpractice()` throws a couple of warnings. it seems a few files don't end in a new line. I'm not sure how consequential that is to functionality. In any case, it probably means the files aren't being parsed correctly during `goodpractice` checks so it's best to address it we can check all the code.

- [ ] ensure complete final lines in files:

Preparing: lintr
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/outsider/R/identities.R'incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/outsider/tests/test-all.R'incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/outsider/tests/test-all.R'incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/outsider/tests/testthat/test-outsider.R'incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/outsider/tests/testthat/test-outsider.R'incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/outsider/tests/testthat/test-test.R'incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/outsider/tests/testthat/test-test.R'
```

Finally an issue that needs addressing

- [ ] remove call to `require()`

  ✖ avoid the library() and require() functions, they change the
    global search path. If you need to use other packages, import them. If you
    need to load them explicitly, then consider loadNamespace() instead, or as
    a last resort, declare them as 'Depends' dependencies.

    R/test.R:18:22

```

Could you also please:

- [ ] add the rOpenSci under review badge to your README?

If you could try and address these before review begins officially, that would be great. I'll ping back in when I've secured reviewers. In the meantime, feel free to reach out with any questions

---
