### 2019-02-18 #282 @DomBennett
Hi @annakrystalli

Thanks for the editor checks. I hope I've managed to correct the issues you raised!

Do you have any advice on running tests across different R versions (oldrel, release and devel) locally? Looks like I'd need to set up multiple virtual environments and install and run tests in each one; a bit of a palava.

Thanks!
Dom
