### 2019-02-18 #282 @annakrystalli
Hi @DomBennett,

Thanks for the swift action!

So we've opened an[ issue in the `dev_guide`](https://github.com/ropensci/dev_guide/issues/132) regarding this (feel free to contribute to the discussion) and have outlined a couple of suggestions.

The main approach to test locally would be to use Rocker Docker images with different tags. Get different version with eg `docker run -v $(pwd):/pkg rocker/verse:devel Rscript "devtools::check('/pkg')"`.

In your case It will be a bit trickier, because the package needs access to Docker itself. But it's doable. If you write a dockerfile that installs the docker client on top of a Rocker image, and mounts the docker socket when running the container, it should work.

Hope this helps and let us know if you have any more questions 😸
