### 2019-02-20 #282 @DomBennett
Thanks @annakrystalli! I'll look into using rocker and keep an eye on the issue in the dev_guide.
