### 2019-03-10 #282 @nuest
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

_Wow - quite a big project you've created there. I'm not sure I can cover everything in this first round of review, so please understand I'm putting in a few questions in the following comments._

Please find the full output of my review tests at https://nuest.github.io/outsider-review/index.nb.html

[Disclaimer: I did not do yet write my own module because of time constraints.]

- I make almost no comments on functions and argument  naming, because I could follow the current naming scheme quite well - I think the author did a good job of following the dev guide
- _Contributor guide is missing_
- the version number `0.0.0` seems curious to me, I suggest to either go with the `.9000` versioning [as described here](http://r-pkgs.had.co.nz/description.html#version) if you think this is a development state, if you think you have a first working version (as the "active" repo status suggests), why not go with `0.1.0`?
- The **statement of need** is present, and not bad at all, but it could be improved by adding a target user group. _Is your goal to hide the complexity of using Docker?_ Is the main goal for users to develop their own modules or use modules prepared by you?
- A container is generally not considered a virtual machine, so I suggest rephrasing several paragraphs in the JOSS paper and README and the vignettes and take the time to introduce the basic terminology. There is a trade off between brevity now, and potentially confusing users later once the go deeper into Docker, or have already some knowledge of Docker.
- **JOSS submission**
  - AFAIK the `paper.md` should go into the base directory.
  - Can you clarify the author mismatch between the package and the JOSS paper? I suggets to double-check the JOSS guidelines about who is an author.
- **Modules**
  - To me "hiding" module-creation/developer related functions using the `.`-Notation is misleading. You normally don't want users of your package (be they normal or power users) to rely on such "internal" functions. _What about using different prefixes?_, e.g. `moduledev_..` for developer functions?
  - Re. module names: Is the `om..` notation really only for discovery on GitHub? If so, have you considered using GitHub tags instead? If you don't require specific names, people could enhance their existing R packages to become otutsider modules.
    - The actual search for packages does not only require the name to start with om but also the description to have "outsider module" - I suggest to document this clearly for module authors - for them being listed on you package page is very important!
  - What do you think about letting the images be built on Docker Hub, instead of building locally and pushing them? May be something for an advanced vignette.
  - `DESCRIPTION` metadata vs. `om.yml`: Can you elaborate on the reasoning to put the Docker user name in the `DESCRIPTION` into a custom field, and not into the `om.yml` ? At first sight it might be easier for users to have the `DESCRIPTION` file as standard as possible and put your own metadata in one place. Or the other way around: Why not put `program` and `details` into the `DESCRIPTION` file?
  - _Is there a way to list all function names in a module?_ The Get Started vignette says "Once a function name is known of a particular .." - can I only do that via the help file? MAybe a `module_list(repo = 'dombennett/om..mafft')` is feasible to implement.
- _Do you have any further development plans?_ I can imagine further developments (like building pipelines using a chain of modules), maybe you can put your ideas into GitHub issues for everyone to learn about them?
- in `README.md`: I recommend to have an empty line after headlines and a space after the `#`/`##` in the source code for readability
- You list "docker" as a system dependency. _Have considered checking for existence of Docker at installation time?_ Really curious here because I also have a Docker-related package. Could be worth discussing in a `sysreqs` issue, and re-check https://ropensci.github.io/dev_guide/building.html#package-dependencies

#### Source code

- Idea: `.module_validate(..)` to validate the formal structure of a module - I now see there is `.module_check` :-) - I suggest to put this "TODO" into a GitHub issue so that potential collaborators see this is an open task, and not have a function that a user cann access with `:::` without any functionality. Or at least add a `- _Is there a way to list all function names in a module?_ The Get Started vignette says "Once a function name is known of a particular .." - can I only do that via the help file? MAybe a `module_list(repo = 'dombennett/om..mafft')` is feasible to implement.
- _Do you have any further development plans?_ I can imagine further developments (like building pipelines using a chain of modules), maybe you can put your ideas into GitHub issues for everyone to learn about them?
- see comment about dev-functions starting with `.`warning("not implemented")`
- I think the package does a good job to distinguish between user facing functions and internal functions (like `test()`)
- formatting is alright
- for `pkgdetails_get` I personally would rely on the package [`desc`](https://cran.r-project.org/package=desc) to do the parsing - you already have a number of imports and this lightweight one should not hurt
- `.module_travis` - your approach means that the function will fail if the repo mooves, or if there is no internet connection... for me it's always a signal to reconsider the approach if you have a constant like a URL within a function body. I think it would be better to manage the "template" file within the codebase of the package.
- `container.R`: your package code probable precedes `stevedore`, but if you find yourself writing code around calling docker with `sys::exec..` then it might be worth the effort to switch; things like [running containers remotely](https://github.com/AntonelliLab/outsider/issues/1) (if that is what you mean by "remote hosting") would come very cheaply then;
- `search.R`:
  - "code you wrote 6 months ago might just have been written by someone else": I suggest to add a warning after `tkn <- NULL` to make it more obvious when requests fail because of a missing token (and not because the token is not valid etc.)
  - `info[['url']] <- paste0('https://github.com/', info[['repo']])` > reuse `gh_url` from the top of the same file
  - `module_exists` is not implemented fully, see suggestion above (warning and/or GitHub issue)

#### Test suite

- good job on using mocks, very educational for me to read a testsuite with mocks; hard for me though to give suggestions to improve, except: It might be worth the tradeoff of having some "integration tests" without any mocks, because with all the mocks it's hard for me to tell if you actually test a relevant amount of the control flow; you can always skip those tests if it is about speed, e.g. skip them unless and environment variable is set, and just let them run on Travis, where time is not that critical. Tests that I think are worth having as **integration tests** are the interaction with the GitHub API in `test-search.R`, the "main" function testing in `test-outsider.R`, `test-log.R` (i.e. run actual workflows, capture the output, and see if the user is (not) shown what she should see). You could also run integration tests with incomplete modules and see if they fail at the expected point.
- the tests for the skeleton could be extended to cover the minimal contents of the module, not just the directory. I think this might be worth the effort towards an indirect formal specification how a module looks like.
- in `test-docker.R`
  - Tests that actually interact with Docker, e.g. using `docker_pull`, should be skipped on CRAN, see `?testthat::skip_on_cran`, or you add your own test function `skip_if_docker_is_not_available()` (later I saw a comment "TODO: add skips if no docker" - go for it :-) ! )
  - the test "docker_build() and docker_img_rm() works"
    - relies on the `img` variable from the previous test - that can quickly break. I suggest to either duplicate `img`, or move it outside of the `test_that(..)` function
    - you should also test that the second execution of `docker_img_rm()` returns `FALSE`
  - In general, **consider checking the state outside of your own functions**, e.g. `expect_warning(expect_equal(attr(system2("docker", args = c("inspect", img), stdout = TRUE, stderr = TRUE), "status"), 1))`, or with the package stevedore (also applies to `test-container.R`)
- in `test-console.R`
  - I suggest to use `expect_output(..)`, e.g. `expect_output(outsider:::cat_line('cat this'), "cat this")` (and also for `celebrate()` and `comfort()`)
- `outsider:::vars_get()` - the function is only for testing, right? Then I think it would be better to add it to the test codebase, e.g. in `test-all.R`
- since you advertise the package as a cross-platform solution, I think you should add **CI tests for Mac and Windows** using Travis and AppVeyor (AppVeyor with Docker also works, feel free to get in touch if you have problems)

- Add `phylotaR` and `ape` to `Suggests` because you use them in vignettes
- The contents links on https://antonellilab.github.io/outsider/articles/advanced_build.html do not work yet/content is missing
- `outsider` covers a pretty complex problem! With the existing examples of one very basic "hello world" and one (to me very) complex analysis pipeline, I see room for a further example that shows a bit more but is not that domain specific, e.g. loading some data from a CSV and running a regression model, saving that to a file. For me as a user it would be great to know _how can I load my data, how call my function, and how access the output data_.
- **Get Started**
  - The code chunk using `input_sequences.fasta` does not work - would it be possible to let a user download (with R code) such a file and actually run that? I was a bit disappointed when I went though all previous steps and then was "stopped" unexpectedly.
  - users often want to know what's going on, and outsider is doing quite a lot of magic - I suggest to mention `log_set` already in the get started documentation.
  - "Uninstalling" - for users new to Docker, you could mention here how to removed the downloaded images.
- **phyologenetic vignette**
  - suggets to add a link to the outsider module on GitHub (https://github.com/DomBennett/om..mafft), potentially even to specific files, to help users explore this more complex module with the help of the vignette
  - is there are particular reason for the repeated `"library(outsider)"` statements in each section?
  - Can you make the output of `module_help(repo = 'dombennett/om..mafft', fname = 'mafft')` more helpful? Right now it only says "Run mafft". I think for the "story" of the vignette it would be much better to actually have some documentation there, e.g. reference papers and describe the function.
  - In section "Alignment" - what should I look for as a user at the end? Can you add some visualisation here or describe what a user could do now with the output of the process?
- I really like the way you have structured the **module writing vignette** and weaved in background and basic information. I think they could be even better if you'd add more links to external resources (e.g. tutorials on writing Dockerfiles, on the specification of the `FROM` instruction in the Docker docs, `WORKDIR`, ..)
  - You could tell users about the `stevedore` package, which allows them to do the buidling of Docker images from R directly.

_Please excuse my brevity at some points, and don't be discouraged at all but the length of the comments. I think the package is a useful approach to packaging workflows and tools in a user friendly way and a great usage of container technology._

It would be great to see this beeing picked up in your lab/your community of practice and see how people interact with each other's work based on outsider modules.
