### 2019-03-23 #282 @hrbrmstr
NOTE: I deliberately didn't read @nuest's review before going through my own package review. I then read @nuest's most excellent/thorough review and removed [most of, hopefully] any duplicate findings unless I felt the need to reinforce some.

>##### For packages co-submitting to JOSS
>
>- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

#### Quick items from the ^^ checklist:

- RE: "JOSS" items: `paper.md` IMO does contain a sufficient "statement of need" even though it's all in the Summary
- RE: "Functionality" : Automated tests passed on macOS 10.14.4 RC / R 3.5.3
- `devtools::test()` and `devtools::run_examples()` all ran fine on macOS 10.14.4 RC / R 3.5.3

#### Quick personal notes that align with or offer an alternative opinion to items @nuest already said very well:

- While I understand it would be a tiresome exercise I'd like to see the Docker machinations be performed with `stevedore`. _This is not a show stopper_ but it would reduce package size and compelexity and perhaps be less confusing to some R folks.
- In an effort to hold fast to the "Quick" part of the section title, I echo a number of items from @nuest under this simple phrase: "i suggest adding preflight checks" (I'll elaborate below).
- I heartily suggest the a fully built out "Advanced…" vignette be created before the package is accepted (more on this below).

#### My review

(again, trying to not duplicate @nuest, but apologies if there are)

First: Bravo! This was pretty darned cool once I got it going.

Second: Definitely check out "whalebrew" (https://github.com/whalebrew/whalebrew) as I think that might be a good project to side-reference somewhere in the `outsider` docs (since it accomplished a similar but not duplicative goal) and also consider how they've structured their concept of "modules" (package): https://github.com/whalebrew/whalebrew-packages. In fact, it was really only when I "AHA!"'d the fact that `outsider` was "whalebrew for R" that I finally grok'd the package.

I shall now stop with the numerical itemizations :-)

I will now suggest something that will likely cause you to hurl something to me if we ever meet at an R conference: consider splitting the package into two separate packages, with the main `outsider` package being for the general, average, "will likely never build a Docker image in their lives but could make excellent use of existing outsider modules" R users with very targeted documentation and functionality and, perhaps "`outsider.devtools`" that focuses on making it as easy as possible for folks to build, test, and deploy `outsider` modules. I say that for a few reasons. First, the function list in the help index is _extremely daunting_ for a large swatch of R users and most of the functions are for module builders. With a more focused, general-user-level `outsider` package you'd likely get more adoption and could focus all the documentation on just using modules. Further, I found the:

- start a new module =>
- build =>
- test =>
- "DOH I FORGOT SOMETHING!" go back and add stuff to the module =>
- rebuild =>
- "When do I re-push to GitHub again?" =>
- uninstall =>
- fight with lazy load database issues (not necessarily an `outsider` problem) =>
- finally have a working module
- "How do I get output from Docker 'functions' again?" =>
- rebuild =>
- finally have a working module =>
- "Did I forget to push it?"

process a bit…unsatisfying.

Speaking of which, to get to ^^ I _actually built a module!!_ : https://github.com/hrbrmstr/om..nmap : but I'm not sure why the one error shows up on `module_details()` since I followed the directions to create the skeleton and push stuff up to the proper locations:

```r
library(outsider)

repo <- "hrbrmstr/om..nmap"

module_search()

module_uninstall(repo)

module_install(repo)

module_details(repo)
## Error in paste(lines, collapse = "\n") :
##   cannot coerce type 'closure' to vector of type 'character'

module_help(repo)

nmap <- outsider::module_import("nmap", "hrbrmstr/om..nmap")

x <- nmap("-v", "-oX", "/working_dir/outputfile.xml", "-sS", "a.b.c.d)

str(x)

xml2::read_xml(x$outputfile.xml)
## {xml_document}
## <nmaprun scanner="nmap" args="nmap -v -oX /working_dir/outputfile.xml -sS a.b.c.d" start="1553359840" startstr="Sat Mar 23 16:50:40 2019" version="7.70" xmloutputversion="1.04">
##  [1] <scaninfo type="syn" protocol="tcp" numservices="1000" services=" ...
##  [2] <verbose level="1"/>
##  [3] <debugging level="0"/>
##  [4] <taskbegin task="Ping Scan" time="1553359840"/>
##  [5] <taskend task="Ping Scan" time="1553359841" extrainfo="1 total ho ...
##  [6] <taskbegin task="Parallel DNS resolution of 1 host." time="155335 ...
##  [7] <taskend task="Parallel DNS resolution of 1 host." time="15533598 ...
##  [8] <taskbegin task="SYN Stealth Scan" time="1553359841"/>
##  [9] <taskend task="SYN Stealth Scan" time="1553359842" extrainfo="100 ...
## [10] <host starttime="1553359840" endtime="1553359842">\n  <status sta ...
## [11] <runstats>\n  <finished time="1553359842" timestr="Sat Mar 23 16: ...
```

If you do look at https://github.com/hrbrmstr/om..nmap you'll see that the markdown generated by the package generates invalid R code (no double quotes around the repo param values).

All of the recently above is likely summarised by "I think module building documentation and tooling could use a little more work, especially bits about getting data back from them" (between man pages and looking at an example of yours I obviously figured it out, but just before I did I almost would have preferred using whalebrew's nmap image and `system2()`). Once I finally worked through making a module I can completely see the utility of this package and think it's a great addition to the R universe.

Regardless of whether it is split into two packages, I would like to see an on-load/attach pre-flight check and/or pre-flight check function(s). For "users" this could check for Docker presence/version and any other items that are essential to the successful operation of the package. For "builders" it could also check for essential/recommended environment variables and/or anything else that you know might make it easier to do work with the package.

I think the functions themselves are properly documented for their functionality and the code style is fine, but I'm curious as to the inconsistent use of `stringr` functions. I'm also generally not a fan of adding a `stringr` dep vs a `stringi` dep, especially if you aren't using any "in stringr-only functions" since `stringr` depends on `stringi` and `stringi` is already a bear of a dependency on linux systems. In fact, I only see two FQ uses of it so I somewhat question the need for it at all. I didn't do a full check for whether it comes along for the ride with the Imports, but I don't think it does. So I'd suggest validating that and either moving to stringi or just using base R string functions (like you do with `sub()`/`gsub()`).

One bit of tension I also had as a module developer was the check "green" travis checks causing a warning message to be displayed. I think a further check could be made to ensure that the package is even being tested at all and then `message()` that vs `warning()` it.

Finally, as noted a few times, this package installed w/o a hitch as did using modules. That is most, most, most impressive for massive amount of complexity this package hides from the user. Very well done!

#### But (Of _course_ there's a "but" :-)

As someone who only puts his stuff on GitHub now to make it easier for all the R folks who are being led astray there by RStudio, (w/apologies to the team) rOpenSci, DataCamp, et al to find his stuff I'm kinda put off by the absolute need for GitHub. I 100% understand that supporting different social coding sites adds complexity but it also breeds diversity and avoids single points of failure. So, please consider not having a hard-dep on GitHub for module discovery.

What might make ^^ easier to deal with (though it requires a re-jiggering of module discovery and expectations) and to make it far easier for users in general is to mimic what whalebrew does (again, https://github.com/whalebrew/whalebrew-packages) and require modules to be under a single repo. You could do this via submodules so the code stays with the owner (like Hugo themes does) but is "registered" under there. Then, you could keep this single repo mirrored across GitLab, GitHub and other free and less harmful-to-humanity sites.

Speaking of modules… I'm already not a fan security-wise of the existing way GitHub is used to brazenly load Windows binary DLLs onto user systmems from GitHub and this goes a step further and puts the entire infrastructure of a containerized OS onto a local system pretty opqauely (yes, so does whalebrew). Great for abstraction, not great for system or analysis integrity. The winbin dance is going away soon and I think I'd feel a bit better about `outsider` modules if the "single repo with subdirs" approach was taken and there were hard-and-fast set of guidelines in place to be accepted into this new "outsider verified registry". Under the umbrella of rOpenSci these could be established in parallel with traditional onboarding practices but I'd rly argue that:

- validation is done that only ssh is used to work with repo contents
- all commits are signed with keys in Keybase
- someone does a manual inspection anytime a local in-dir copy is updated or a submodule reference hash is changed
- outsider modules must be tied to releases (GL, BB, GH, etc all support the contept of a release) and pass similar checks
- Docker images are also validated somehow

Unlike `apt get` / `brew install` Docker is the "Wild, Wild West" and GitHub is "Max Max". With `outsider` coming under the umbrella of rOpenSci I rly think the need for a more formal module registry is needed since — while it would not be rOpenSci's fault if something bad happened — an integrity or security incident could negatively impact rOpenSci's reputation. Unlike `stevedore` (where it's just a Docker API wrapper) this is very R specific and is also suggesting that this `outsider` approach is sound/safe for scientific computing (and I'm not saying that it can't be). So, to me, that kinda means there should be at least some type of badge/star/etc that goes next to "official/reviewd/clean/supported/safe/secure" `outsider` modules and ones that are just "out there". Perhaps I'm overthinking this and just paranoid (which is a part of the job description). So I'd def like all folks to weigh in here. I will also gladly craft a malicious module if it comes down to that :-)

#### FIN

I'm going to re-visit the package Sunday from scratch with fresh eyes to make sure I didn't miss anything not covered by @annakrystalli or @nuest.

Overall: amazing job; very useful concept and package and the execution is solid as-is tech-wise.
