### 2019-03-26 #282 @DomBennett
Thanks @hrbrmstr for the super thorough, well-thought and detailed review! I've got myself lots of things to be thinking about and getting on with.

Thanks @annakrystalli, I'll get to addressing these issues soon as.

_Merci à toutes et tous_
