### 2019-03-31 #282 @richfitz
Just been lead to this via twitter - if you are interested in using `stevedore` but hit any hurdles @DomBennett, please feel free to contact me on the rOpenSci slack
