### 2019-04-17 AntonelliLab/outsider#3 @DomBennett
outsider.devtools
Create new package of just "builder" related functions and objects.
