### 2019-04-17 AntonelliLab/outsider#4 @DomBennett
Switch to `stevedore`
Switch from system docker calls to `stevedore`.

Just for dev but also for `outsider`-proper?
