### 2019-04-17 AntonelliLab/outsider#5 @DomBennett
Less GitHub dependence
Add compatibility for other code-sharing sites.
