### 2019-04-17 ropensci/outsider#6 @DomBennett
AppVeyor
CI for Mac and Windows
