### 2019-04-17 AntonelliLab/outsider#7 @DomBennett
Testing suite: interactions and pipelines
Testing should cover wholesale interactions. Create mock modules (both broken and functional) to enrich testing process.
