### 2019-04-17 AntonelliLab/outsider#8 @DomBennett
Documentation, documentation, documentation
- advanced vignette for builders
- more pipeline examples, more vignette examples
- embellish phylogenetic vignette storyline
