### 2019-04-17 AntonelliLab/outsider#9 @DomBennett
Add future directions issues
- create future directions issues
- describe in detail (with figures!) my dreams for outsider
- create a new label to identify them
