### 2019-04-17 AntonelliLab/outsider#10 @DomBennett
Security and module discoverability
- How to make modules both discoverable but also ensure they're not malicious?
- Creating a "approved" list of modules may be difficulty to maintain and may lead to problems if a module is misidentified
- Is a simple warning system enough?
