### 2019-04-17 #282 @DomBennett
Hej allihopa!

Once again, thanks for the super-duper detailed and insightful reviews to the package. Lots of good advice and angles I hadn’t considered. Sorry for my lateness getting to this; I was juggling many things, plus I always have trouble approaching reviews for the first time. 😬

Given the complexity of this package and the high-level suggestions provided by the reviewers, I figured I should aim to be more communicative than in my previous submissions to ROpenSci where I just implemented the changes and dumped a great big message in the review. Here, I think a bit more to and fro is required.

So to start off with, I will outline below both the major and minor changes I intend to implement as well as the changes I would like a bit more discussion/help with before I implement anything. As I implement changes, I will then provide sit-reps on progress and setbacks.

## Major changes I will make

### `outsider.devtools`

As both reviewers indicated, the mixing of dev and user functions is confusing. I actually have no issue in creating a separate package and siphoning off the strictly dev related tools to this new package. I actually like this idea and think it is the cleanest approach. I assume ROpenSci would have no quibble in hosting the additional package.

Issue: https://github.com/AntonelliLab/outsider/issues/3

### Switch to `stevedore`

I will definitely switch code to `stevedore` for `outsider.devtools`. This makes a lots of sense to reduce community-level code duplication and to support future features and docker dev within R. I am, however, a bit tentative about implementing `stevedore` for `outsider`-proper because of the potential issues for Windows users (https://github.com/richfitz/stevedore#windows-support). (@richfitz, would Windows users be unable to pull and run docker images without the need to install all the dependencies that you list?)

Issue: https://github.com/AntonelliLab/outsider/issues/4

### Less GitHub dependence

@hrbrmstr, agreed. Other options should be available. It should be possible to create tools that can either search GitLab, BitBucket or whatever else. As an absolute failsafe, it should be possible to install a module from a local directory.

Issue: https://github.com/AntonelliLab/outsider/issues/5

###  AppVeyor

Implement AppVeyor CI. This is a major issue for me as I’ve never used Appveyor! If I have any issues, I’ll let you, @nuest, know. I’m just going to see others do it and mimic.

Issue: https://github.com/AntonelliLab/outsider/issues/6

### Testing suite: interactions and pipelines

@nuest, I actually went for the mocks because I was having big issues with the time the tests were taking and I was having issues with uninstalling and reinstalling packages, as part of the module tests, on Travis. There were issues related to lib paths that I tried to fix using `withr::with_temp_libpaths` but only had mixed success.

I can solve the time issue by using skips. I also like testing for functional and non-functional modules. For these larger tests of interaction, I may encounter issues with the libpaths again. I’ll will try and fix these issues.

Issue: https://github.com/AntonelliLab/outsider/issues/7

### Documentation, documentation, documentation

I will splice the documentation between `outsider` and `outsider.devtools`. I will also try and create a middle of the road vignette that is not technical (biological sequence alignment) nor too simple (hello world). The advanced vignette will move to `outsider.devtools` and be enriched with all the `stevedore` goodies with extra links to pre-existing documentation.

Issue: https://github.com/AntonelliLab/outsider/issues/8

###  Issues for future dev

I have a few ideas for the future direction of `outsider`, many for which I have no idea how to implement. @nuest, I will do as you say and create them as separate issues.

Issue: https://github.com/AntonelliLab/outsider/issues/9

## Major changes I need guidance with

### Security and module discoverability

@nuest suggests I drop the need for “om..”. I can do this. I actually already make use of the GitHub description in addition to the “om..” and the presence of `om.yml`. I think I did these three tests initially because I thought the more tests there are for determining whether a repo was a module or not the better. In hindsight, this is probably not necessary. It may be better to just use the GitHub labelling system. A label like “outsider-module” should work just fine. (But, then, it will be a different system on other code-sharing sites.)

@hrbrmstr suggests I add all “approved” modules to a single repo and that I might be able to make use of GitHub’s `submodules` feature. I looked into `submodules` early on in `outsider`’s development and found that they were not universally praised by all those who used them. More simply, I could just have a list of approved repo URLs in the main `outsider` repo. (This might allow for non-GitHub modules as well.)

Ultimately, however, I didn’t want to create lots of work for myself nor others – such as those at ROpenSci who would, presumably, be hosts for this “approved” modules listing. I understand the need to ensure that modules are not malicious, but I do worry that trying to police them may be a task too great to take on. Additionally, are we even able to determine whether a module is safe or not? I might be able to determine the non-malicious status of a module in my own field of research (biodiversity, macroevolution, phylogenetics) because I’d likely know the people and recognise the software being run. But I’m not sure if I could outside of my field; we researchers tend to write quite obfuscated code by default. The worst-case scenario might then be a module is approved, which then turns out to be malicious. Would I be held responsible for any potential fallout? Also, do we have any suggestions on what sorts of malicious things could be achieved through an `outsider` module? And, are there anyways of ensuring they’re not doing these sorts of things?

My thinking to save community-effort is to allow users to download any module they like from whichever code-sharing site but to provide warnings along the lines of:

```
Beware: outsider modules have the potential of introducing malicious software.

This repo has # many watchers and # many stars. The maintainer of the repo is #.
Please make sure you are confident this is a legitimate outsider module.
Note, outsider maintainers take no responsibility for code developed by external developers.

Are you sure you wish to proceed?
(Y/N)
```

Perhaps this warning could appear on any code that has too few stars/watchers?

(Presumably I’m being paranoid now considering that `whalebrew` has an approved list. Plus this all assumes that `outsider` gets picked up.)

Issue: https://github.com/AntonelliLab/outsider/issues/10

## Minor changes that will be made

- Add contributor guide
- Change version number to 0.1.0
- Paper updates: statement of need, author clarifications, information clarification, move `paper.md`,
- Update `om.yml`, allowing: dropping `om..` requirement, removal of metadata from DESCRIPTION.
- List functions in a module
- README spaces
- Docker as a system requirement: can it be auto-installed?
- Create a `module_check/validate`
- Use package `desc`
- `.module_travis`: don’t make dependent on URL
- Fix `search.R` issues
- Tests: img, skips, use `expect_output` (I didn’t know this existed!), move strict testing functions to `test-all.R`
- Suggests: phylotaR, ape
- More story in phylogenetics vignette.
- Add more references to whalebrew
- Pre-flight checks: user and build checks to flag potential issues
- Drop highly suspect use of `stringr`
- Use message instead of warning for module testing
- Have a play with @hrbrmstr’s module

Issue: https://github.com/AntonelliLab/outsider/issues/11

##  Additional points and questions

@annakrystalli, there’s a ROpenSci slack? Would it be possible to have access to it?

Thanks!
