### 2019-04-25 #282 @annakrystalli
Geia sou @DomBennett! 😜

Many thanks for your detailed response and apologies for the delay in mine!

I see the invitation to the slack has gone through so welcome! Feel free to ask questions for broader feedback there.

Re:
> As both reviewers indicated, the mixing of dev and user functions is confusing. I actually have no issue in creating a separate package and siphoning off the strictly dev related tools to this new package. I actually like this idea and think it is the cleanest approach. I assume ROpenSci would have no quibble in hosting the additional package.

Yes that should be fine, it would be deemed in scope.  We'll need to check that:
- [ ] documentation was ported appropriately as you suggest,
- [ ] The relevant tests are also ported and checks and tests all pass,
- [ ] a README that allows the package to standalone is included.

If all these checks pass, the package should also be considered reviewed.

Re: Your security concerns, I like the approach you've proposed. Let's wait to hear reviewers' opinions.

Let me know if you need any more info from me. Looking forward to @nuest & @hrbrmstr input!
