### 2019-05-09 #282 @DomBennett
Hallå hallå,

**Quick update on progress**

I’ve actually decided to split the package three ways: [`outsider.base`](https://github.com/AntonelliLab/outsider.base), `outsider` and [`outsider.devtools`](https://github.com/AntonelliLab/outsider.devtools/).

Why? I actually found it quite hard to determine what should be exported objects if I just had `outsider` and `outsider.devtools`. But now there’s quite a clear line of dependence and clear statements of purpose of each package.

![package_structure](https://raw.githubusercontent.com/AntonelliLab/outsider.devtools/master/other/package_structures.png)

**Other updates**

* `outsider` and `outsider.base`, along with all the previous modules I made, are now functional (as tested so far, at least).
* `outsider` can install modules from GitHub, GitLab, Bitbucket, URL or downloaded file path.
* I have kept the same process of module discovery (only implemented for GitHub ATM) and have added a big health warning.

**Still TODO**

* `outsider.devtools`
* Richer test suites
* AppVeyor
* Future dev ideas
* Documentation
* A healthy plethora of other things
