### 2019-05-24 #282 @nuest
Hi all, just catching up on the many good plans. Looks very promising and a huge step forward for the package(s). Here are some answers/comments:

- Re. "Security": I agree with your concerns regarding policing and managing of such a list of modules. In the end it boils down to "Do I know the person and trust her?", potentially judging that via online reputation. Using stars or watchers will only break at some point (if I am evil enough to try to get my code onto your machine, do you think I won't be able to fake watchers and stars?).
  - I think the best way to go forward is to add clear warnings to vignettes, and have that warning in the code that you outlined. And then, yes, hope that `outsider` get's so much attention that the first security issue arises! Be aware, educate users, but don't fix problems that don't exist.
  - Idea: Allow users to switch off the warning by adding a variable in their R profile?
  - I have not checked, but you could add a "Security" section in a vignette, and point to the basics of container security, and suggest users review the code of modules they install.
  - _What is really the worst that could happen?_ I'm really asking! The malicious aspects I can imagine is that someone might steal computing power (you could add an option to disable network access for the started containers so they cannot call home) or that a malicious file ends up on my machine (which you can warn users to be careful in the security statements). What else?
- Re. "Module discoverability": Unless you're willing/able to maintain a single point of service yourself forever, you have to live with having different means to "tag" a repo for each supported service. But you tool is open source, so if you support GitHub and GitLab, others can contribute more services. I'd still just drop `om..+in:name` from you GH search string :-).
- Re. `outsider.base`: since `outsider.devtools` devtools will always rely on `outsider` (right?), I think factoring out the common base was not a strict requirement. But small R packages are getting more common, so :+1:

------

**Comments on code changes**

(I browsed through the code changes only.)

- `tests/testthat/test-download.R`: testing `url_make()`, why do you not test the content of the URL but only the type?
- nice use of `remotes` to install from the different sources! Have you considered supporting `remotes::install_git(..)` as well? Many research institutes run their own GitLab instances these days and a plain git link should always work then. If you want to save some work, maybe add to the documentation that the regular URL also works for any third party/self-hosted git collaboration site.
- in general in your tests, I suggest to check for a bit more than just type or inheritance, but at least on "content" part (cf. `test-github.R`)

------

I applaud your efforts in getting things really right, but have to admit given the amount of things to do I am a bit overwhelmed and have a hard time giving feedback. I think you're on the right track! If you need some specific feedback, please feel free to mention me in GitHub issues, and if I don't get back to you within a couple of days, ping me via email or [Gitter](https://gitter.im/nuest).
