### 2019-05-24 #282 @DomBennett
Thanks @nuest  -- I will act on your early feedback here!

On a general point to the reviewers, please don't feel overwhelmed or obliged to respond to _all_ the points I raise in this thread. I just want to lay out my thinking and show the evolution of the package(s) in an open and accessible way.
