### 2019-05-28 #282 @annakrystalli
Thank you for the feedback @nuest! ✨ And please don't worry @DomBennett.

I can somewhat understand where @nuest is coming from because the splitting up of the package does add additional checking that perhaps wasn't envisaged at the start. For completing the review, I would refrain from tackling "future dev ideas" and try and limit the "A healthy plethora of other things" on your TODO list to changes that are strictly necessary for functionality or are responding directly to reviewers comments. (Although by all means, track them as issues in the repositories themselves). As I'm sure you know, development can and always does, continue beyond review, in response to a broader user base (which may even lend a hand too!).

But just to dispell any worries you or @nuest have and given that @hrbrmstr might not be available to continue, I will be happy to do the final checks to help spread the load a bit. So carry on @DomBennett and just let us know when you think you're ready for final checks. 👍
