### 2019-08-22 AntonelliLab/outsider#6 @DomBennett
Fixed. https://ci.appveyor.com/project/DomBennett/outsider

But! I can't get linux based containers working.

Docker comes pre-installed when using the "Windows Server 2019" image but only with Windows-based containers. (`docker run hello-world` doesn't work.)

May have to set-up docker with VM as part of the build script. Perhaps using [`chocolatey`](https://chocolatey.org/)?

(Issue part of [ROpenSci review](https://github.com/ropensci/software-review/issues/282))
