### 2019-08-22 AntonelliLab/outsider#4 @DomBennett
Problems with implementing change:
- `stevedore` does not run on Windows (easily)
- would require a lot of rework
- could be done at a later stage without changing user interaction
