### 2019-08-22 AntonelliLab/outsider#10 @DomBennett
- All modules are installed with a health warning
- Modules are discoverable by a simple naming convention (`om..`, "outsider" label)
- Modules can be advertised through other means, there is no requirement for modules to be strictly light-weight and following the naming conventions. They can be whole R packages. They just won't be automatically discoverable.
