### 2019-08-22 AntonelliLab/outsider#11 @DomBennett
- [ ] Add contributor guide
- [x] Change version number to 0.1.0
- [ ] Paper updates: statement of need, author clarifications, information clarification, move paper.md,
- [x] Update om.yml, allowing: dropping om.. requirement, removal of metadata from DESCRIPTION.
- [ ] List functions in a module
- [ ] README spaces
- [ ] Docker as a system requirement: can it be auto-installed?
- [x] Create a module_check/validate
- [ ] Use package desc
- [ ] .module_travis: don’t make dependent on URL
- [x] Fix search.R issues
- [ ] Tests: img, skips, use expect_output (I didn’t know this existed!), move strict testing functions to test-all.R
- [ ] Suggests: phylotaR, ape
- [ ] More story in phylogenetics vignette.
- [ ] Add more references to whalebrew
- [ ] Pre-flight checks: user and build checks to flag potential issues
- [x] Drop highly suspect usage of stringr
- [ ] Use message instead of warning for module testing
- [ ] Have a play with https://github.com/hrbrmstr/om..nmap
