### 2019-08-22 AntonelliLab/outsider#6 @DomBennett
Clarification: `docker` is only required for `outsider.base`

AppVeyor project: https://ci.appveyor.com/project/DomBennett/outsider-base
appveyor.yml: https://github.com/AntonelliLab/outsider.base/blob/master/appveyor.yml
