### 2019-08-26 #282 @DomBennett
Sorry for my massively slow updates: lots of things were happening, this project is way bigger than any of my other ROpenSci projects.

## Responses to the reviews

* As already mentioned, the project is split across three packages now
* The base package interacts with docker, [link](https://github.com/AntonelliLab/outsider.base)
* The main package provides nice wrapper functions for users and allows for module discovery and installation, [link](https://github.com/AntonelliLab/outsider)
* The dev package provides helper functions for creating one’s own module, [link](https://github.com/AntonelliLab/outsider.devtools)
* Each package is documented and tested with good coverage (>90%)
* AppVeyor is there for all packages
* Users can install modules via github, gitlab, bitbucket, url, file path or git repo
* Users are provided with a security warning whenever they install a module.
* Lots of other minor changes as per the reviewer comments (version number, module checking, list a module’s functions … etc. see [issue 11](https://github.com/AntonelliLab/outsider/issues/11))

## One new core feature

* ssh: launch docker from a server

## Small things still lacking/fixing

* [ ] Documentation: update basic module build example
* [ ] Documentation: more complex module building example
* [ ] Documentation: example of using `ssh` feature
* [ ] Documentation: future directions
* [ ] Collaborating.md
* [ ] Richer tests suites to be implemented
* [ ] AppVeyor cannot run docker, see [issue 6](https://github.com/AntonelliLab/outsider/issues/6)
* [ ] paper.md needs moving and updating
* [ ] More detailed security warning upon module install: get details from URL

## Notes

* *Module discoverability*: I’ve kept the requirement for modules to have “om..” at the beginning of their names. This means I can search across GitHub, GitLab and BitBucket. “om.. ” is only required to find modules across these services, “om..” is not required for modules to run. There is the additional requirement in GitHub for a repo to have the “outsider-module” tag. Tags are not available in GitLab and BitBucket. Once candidate modules are found, they are then checked to see that they look like modules (R, om.yml).
* *Three packages*: I realise that there is a common denominator. But I like how having the three allows me to control what functions are available to different user groups. For the person who wants to just run modules, calling `outsider` keeps all of the `outsider.base` functions hidden – avoiding confusion. For the developer, they need to learn the exported `outsider.base` functions, upon calling the `outsider.devtools` these functions are made public.
* *Things not done*: `stevedore` has not been used, `desc` has not been used. These packages appear useful but since I’ve already coded the base without them I see no reason in switching to using them just not. If new features are required or errors can be fixed with these packages, they can always be added at a later stage without impacting the user experience.
* *Auto-installing docker*: I will look into this by lodging an issue with [`sysreqsdb`](https://github.com/r-hub/sysreqsdb/issues). It should be possible for linux machines but would require extra scripts, not directly available via any packaging suites (apt, RPM). Plus then there is the question of docker flavour (ce or ee?). I know too little.

I hope these changes are going in the right direction. Let me know of any feedback or issues you have.
