### 2019-09-23 #282 @annakrystalli
This is great work!

@nuest or @hrbrmstr do you perhaps have time to comment on some of the changes made?
I'm happy to do the final checks on all three packages when all the minor fixes still pending are addressed, but your higher-level input at this point on the decisions being made could be really useful.

@DomBennett , should we decide on the minor tasks that need completing to be able to draw the line when they are done? Are you happy to complete the checklist you've got in your issue or are there some that you consider enhancements and therefore not necessarily critical to the issues raised by reviewers?

* [ ] Documentation: update basic module build example
* [ ] Documentation: more complex module building example
* [ ] Documentation: example of using `ssh` feature
* [ ] Documentation: future directions
* [ ] Collaborating.md
* [ ] Richer tests suites to be implemented
* [ ] AppVeyor cannot run docker, see [issue 6](https://github.com/AntonelliLab/outsider/issues/6)
* [ ] paper.md needs moving and updating
* [ ] More detailed security warning upon module install: get details from URL

Once we establish what's left to complete, it would be good to get an idea of a time frame in which they might be complete. 👍

Sound good?
