### 2019-10-25 AntonelliLab/outsider#11 @DomBennett
Issues dealt with
