### 2019-10-25 AntonelliLab/outsider#6 @DomBennett
Issue raised here: https://github.com/krlmlr/r-appveyor/issues/168

---

Should be fixed with time.
