### 2019-10-25 AntonelliLab/outsider#7 @DomBennett
Done: https://github.com/AntonelliLab/outsider-testsuites
