### 2019-10-25 AntonelliLab/outsider#9 @DomBennett
Dealt with in `CONTRIBUTING.md`
