### 2019-10-25 #282 @DomBennett
Hi @annakrystalli !

Thanks for your response. I've been quietly beavering away at the list. I have addressed all of the issues now with one or two cavets:

* Documentation: update basic module build example ([Done](https://antonellilab.github.io/outsider.devtools/articles/intermediate.html))
* Documentation: more complex module building example ([Done](https://antonellilab.github.io/outsider.devtools/articles/advanced.html))
* Documentation: example of using ssh feature ([Done](https://antonellilab.github.io/outsider/articles/ssh.html))
* Documentation: future directions (Considered this to be the same as the next point ....)
* Collaborating.md ([Done](https://github.com/AntonelliLab/outsider/blob/master/CONTRIBUTING.md))
* Richer tests suites to be implemented ([Done](https://github.com/AntonelliLab/outsider-testsuites))
* AppVeyor cannot run docker, see issue 6 (Not done... but new [issue raised](https://github.com/krlmlr/r-appveyor/issues/168))
* paper.md needs moving and updating (Not moved, don't think is necessary, but [updated](https://github.com/AntonelliLab/outsider/blob/master/paper/paper.md))
* More detailed security warning upon module install: get details from URL (Done)

(Additionally, there are come Appveyor issues to do with installation of packages for the devel R release. And there seems to be an issue with installation of `fs` on Travis for my oldrel.)

How should we proceed next?

Thanks!
