### 2019-10-27 #282 @annakrystalli
Great work @DomBennett !!

Well done for chipping away. It sounds like you are in a place where we could potentially start wrapping the review up! All 3 packages need to be looked at though. As I will be doing it, I'm planning to take next weekend to spend some time on it.

I'll also have a look at and enquire about the issues you are having with testing on `devel` and `oldrel`.   In the meantime, I wonder if this might help: https://github.com/r-lib/fs/issues/90
