### 2019-10-29 #282 @DomBennett
Thanks @annakrystalli !

In terms of the CI issues, I tried fixing the failing Travis build for `oldrel` for the `outsider` package based on that issue you directed me to. I tried installing `fs` by specifying it as a binary package, this prevented the `fs` error but then I ended up with a new installation error for a different package. So I figured this is clearly a general compilation issue on travis... which led me to this issue: https://github.com/wesm/feather/issues/154

I followed the [travis YAML guidance](https://github.com/AntonelliLab/outsider/blob/bf9cb7a0f121ac842a5c607d2112507d0bb347f5/.travis.yml) specified in the above issue and found that .... it didn't fix the problem, same old installation error with `fs`.

So for now, I've dropped `oldrel` from the `outsider/.travis.yml`. I can try it out again later. I figure at some point the error would be fixed regardless of my efforts.

One puzzle: why should this be an issue for `outsider` but not `outsider.devtools`? They have virtually the same `.travis.yml`s (Or at least, `outsider.devtools` has all the same dependencies.)

* https://github.com/AntonelliLab/outsider/blob/master/.travis.yml
* https://github.com/AntonelliLab/outsider.devtools/blob/master/.travis.yml

 :confused:
