### 2019-11-25 #282 @annakrystalli
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

## Review Comments

Thanks for your patience with the review. I've completed the review for `outsider`. Will get back to you with the review on the other two packages later today.

Firstly well done on the great work you've done on the package. It feels like the splits you decided on work well and it seems thhe functionality works well. There are still however a few issues to work through, not suprising given the extensive refactoring you've done. So let's work through them:

Running
```r
devtools::install_github("AntonelliLab/outsider", dependencies = T, build_vignettes = T)
```
on source code returns the following error

Running /Library/Frameworks/R.framework/Resources/bin/R CMD \
  INSTALL \
  /var/folders/8p/87cqdx2s34vfvcgh04l6z72w0000gn/T//RtmptOYBoM/outsider_0.1.0.tar.gz \
  --install-tests
* installing to library ‘/Library/Frameworks/R.framework/Versions/3.6/Resources/library’
ERROR: dependency ‘outsider.base’ is not available for package ‘outsider’
* removing ‘/Library/Frameworks/R.framework/Versions/3.6/Resources/library/outsider’
```

- [ ] The `outsider.base` package remote needs referencing in the `DESCRIPTION` file like so:
      AntonelliLab/outsider.base
  ```
- [ ] any other DESCRIPTION files (ie in other packages and your examples), will need remotes specified too.

### Installation problems with `outsider.base`

However, there is a problem with installation of `outsider.base` from GitHub still, with this error:

Downloading GitHub repo AntonelliLab/outsider.base@master
'/usr/bin/git' clone --depth 1 --no-hardlinks --recurse-submodules https://github.com/DomBennett/om..hello.world /var/folders/8p/87cqdx2s34vfvcgh04l6z72w0000gn/T//RtmpZ8wDI8/remotes14452732bf805/AntonelliLab-outsider.base-b8708e3/om..hello.world
dyld: Library not loaded: @rpath/libswiftCore.dylib
  Referenced from: /Applications/Xcode.app/Contents/Developer/usr/bin/xcodebuild
  Reason: image not found
git: error: unable to locate xcodebuild, please make sure the path to the Xcode folder is set correctly!
git: error: You can set the path to the Xcode folder using /usr/bin/xcode-select -switch
Error: Failed to install 'outsider.base' from GitHub:
  Command failed (71)
In system(full, intern = TRUE, ignore.stderr = quiet) :
  running command ''/usr/bin/git' clone --depth 1 --no-hardlinks --recurse-submodules https://github.com/DomBennett/om..hello.world /var/folders/8p/87cqdx2s34vfvcgh04l6z72w0000gn/T//RtmpZ8wDI8/remotes14452732bf805/AntonelliLab-outsider.base-b8708e3/om..hello.world' had status 71
  ```
I solved it by running:

sudo xcode-select --switch /Library/Developer/CommandLineTools/
```
So all good but, given the heavy system requirents, it might be good to consider collecting common troubleshooting advice in a vignette like you have in `outsider.devtools` docs.

### `check()`

There is a couple of test failures for me. It seems the gitlab api call is returning unauthorised.

- [ ]  Are there API tokens required to test any of the functions? If so, include details of what is needed and how to get the API keys in `CONTRIBUTIONG.md`

```
── Test failures ───────────────────────────────────── test-all ────

> library(testthat)
> test_check("outsider")
Loading required package: outsider
----------------
outsider v 0.1.0
----------------
- Security notice: be sure of which modules you install
── 1. Error: gitlab_repo_search() works (@test-gitlab.R#4)  ────────
HTTP error 401.
 1. outsider:::gitlab_repo_search(repo = "DomBennett/om..hello.world")
 2. jsonlite::fromJSON(search_url)
 3. jsonlite:::parse_and_simplify(...)
 4. jsonlite:::parseJSON(txt, bigint_as_char)
 5. jsonlite:::parse_con(txt, bigint_as_char)
 7. base::open.connection(con, "rb")

── 2. Error: gitlab_search() works (@test-gitlab.R#10)  ────────────
HTTP error 401.
 1. outsider:::gitlab_search()
 2. jsonlite::fromJSON(search_url)
 3. jsonlite:::parse_and_simplify(...)
 4. jsonlite:::parseJSON(txt, bigint_as_char)
 5. jsonlite:::parse_con(txt, bigint_as_char)
 7. base::open.connection(con, "rb")

══ testthat results  ═══════════════════════════════════════════════
[ OK: 56 | SKIPPED: 0 | WARNINGS: 3 | FAILED: 2 ]
1. Error: gitlab_repo_search() works (@test-gitlab.R#4)
2. Error: gitlab_search() works (@test-gitlab.R#10)

Error: testthat unit tests failed

1 error ✖ | 0 warnings ✔ | 1 note ✖
```

I also get a note:

```r
  Non-standard file/directory found at top level:
    ‘CONTRIBUTING.md’
```
- [ ] Just use `use_build_ignore("CONTRIBUTING.md")` to add the file to `.Rbuildignore`.

### `goodpractice()`

A couple of super minor points thrown up.

```r
  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your lines

    R/gitlab.R:64:1

  ✖ not import packages as a whole, as this can
    cause name clashes between the imported packages.
    Instead, import only the specific functions you need.
```
I'm tempted to say that because you're importing outsider.base that might be ok. In any case you could consider whether you indeed need the whole package imported? [This](https://community.rstudio.com/t/re-export-all-objects-from-a-package/4295/4) might help with your consideration.

### `spell()`

Only one valid typo I could find! Awesome!

```
  WORD            FOUND IN
Analsis         phylogenetic_pipeline.Rmd:104

```

- [ ] image `alignment_example.gif` on pkgdown site landing page missing.
- [ ] would be nice to have the github link in the pkgdown docs navbar.
- [ ] `is_module_installed` documentation confusing. It seems to be a copy of `module_uninstall` docs.
- [ ] `Details` entry in `module_import` reads: `If program is successfully removed from your system, TRUE is returned else FALSE`, is that right?.
- [ ] One thing to consider in the `Get Started` article is that users not hugely familiar with the command line might benefit from being walked though what each argument being passed is doing is `mafft(arglist = c('--auto', 'input_sequences.fasta', '>',
                  'output_alignment.fasta'))`
- [ ] Given that security was highlighted as an important topic in the reviews, I'm wondering whether some more explicit mentions of security concerns are included somewhere relatively prominently in the docs. You could start by summarising the concerns raised in the review and use it to collate further information as required. This could be in the README or a dedicated article. What do you think?
- [ ] In general you do not need to provide docs for private functions. It could be confusing for users as they can't actually use them unless the use `outsider:::`. You can leave them if you prefer but I would generally recommend focusing on extensive docs for user facing functions.
- [ ] It's best to include a code of conduct alng with `CONTRIBUTING.md`. See `usethhis::use_code_of_conduct()`.

#### `om..*` module docs.

Module repos need a bit more documentation.

- [ ] README need more detail. Someone arriving at the repo should be able to get atleast a high level understanding of what the repo contains. Perhaps you could develop a README template with basic info (perhaps even a function in `outsider.devtools` a bit like `usethis::use_readme_rmd()` to create them). You can that you customise for each module. with some brief info on what the module does and how to use an example function, including relevant code, much like any other package README.

- [ ] It would be good in the module function documentation to either define the parameters or link to official documentation for the module/function and it's parameters. ALthough `fname('--help')` can be enough, sometimes it's not, eg. `bamm('--help')` does not yield much info.
- [ ] Function descriptions also need a bit more info on what they do, eg in `om..bamm`, here should be a short description of what BAMM is. The `details` field (eg `Bayesian Analysis of Macroevolutionary Mixtures`) returned as part of `module_details()` would in most cases suffice, ideally with links to further information and references.

I won't hold up completing the review for this but I encourage you to both take the time to complete it for your modules and consider what best practice you want to model for users and how you can make it easy for them to follow said practice. This could have a large impact on usability of the modules produced by users.
