### 2019-11-25 #282 @annakrystalli
I've completed most checks but want to spend a bit more time working with the package so back again tomorrow.
