### 2020-01-08 #282 @annakrystalli
Great work! And many apologies for dragging my heels with finishing the reviews. Will have it with you by the end of the weekend. Thanks for your patience!!

Sent from my iPhone

> On 7 Jan 2020, at 15:57, Dom Bennett <notifications@github.com> wrote:
>
> ﻿
> Thanks very much @annakrystalli! Super helpful.
>
> I've got through and made the following changes:
>
> Added a remotes section to DESCRIPTION
> Ditto for outsider.devtools
> Added Xcode and Rtools information to the "Getting started" page
> I added a skip to the GitLab API checks if no token is found. I've added information about access tokens to CONTRIBUTING and the documenation.
> CONTRIBUTING is now being ignored
> GP checks: No lines are too long now
> GP checks: I'm going to suggest ignoring the "do not import an entire package" because outsider.base was designed to be used like this.
> Spelling, fixed!
> Documentation: Alignment GIF, fixed!
> Documentation: GH link, added!
> Documentation: is_module_installed, fixed!
> Documentation: `module_imports, fixed!
> Documentation: more info on command-line, added!
> Documentation: more info on security, added!
> Documentation: I've removed "private functions" from the reference on the website
> Code of conduct, added!
> om.. docs
>
> I'm going to work on your suggesed changes to these next. Will work on them in conjunction with updating outsider.devtools.
>
> Thanks!
>
