### 2020-01-09 ropensci/outsider#14 @DomBennett
Update documentation
Need to update documentation for the following modules.

* [x] https://github.com/DomBennett/om..trimal
* [x] https://github.com/DomBennett/om..pasta
* [x] https://github.com/DomBennett/om..partitionfinder
* [x] https://github.com/DomBennett/om..mafft
* [x] https://github.com/DomBennett/om..figlet
* [ ] https://github.com/DomBennett/om..blast
* [ ] https://github.com/DomBennett/om..pyrate
* [ ] https://github.com/DomBennett/om..revbayes
* [ ] https://github.com/DomBennett/om..beast
