### 2020-01-09 #282 @DomBennett
Hi @annakrystalli!

As a quick update, I've updated the documentation for [om..bamm](https://github.com/DomBennett/om..bamm) plus a few other modules ([om..astral](https://github.com/DomBennett/om..astral), [om..raxml](https://github.com/DomBennett/om..raxml), [om..wikit](https://github.com/DomBennett/om..wikit), [om..pathd8](https://github.com/DomBennett/om..pathd8)).

I've also created a template `README.Rmd` as part of the module skeleton for `outsider.devtools`, https://github.com/AntonelliLab/outsider.devtools/blob/master/inst/extdata/template_README.Rmd

The new template suggests developers provide a detailed example with output, information on some key arguments and links and references to the original program.

I've not had time to update the documentation for all the current modules that I maintain. But I've created an issue to remind that it needs done: [Issue - Update documentation](https://github.com/AntonelliLab/outsider/issues/14)
