### 2020-01-13 #282 @annakrystalli
Awesome work @DomBennett ! I really like where these packages are going. It's been  long journey so well done for sticking with it!

OK so here's the review for `outsider.devtools`. Generally only minor points, primarily on the documentation. I also realise that because `outsider.base` is not outward facing, it does not need a formal review.  So after responding to the `outsider.devtools` review points and a quick double check, I believe we will be good to go! 😃🙌

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

## Review Comments

### checks and tests

All good! :+1:

### `goodpractice::gp()`
- [ ] Super minor

```r
── GP outsider.devtools ────────────────────────────────────────────

  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows
    that are about 80 characters wide. Try make your lines

    R/utils.R:87:1
────────────────────────────────────────────────────────────────────
```

### Spell check

- [ ] Only a few potential valid typos!
  ```
  memmory     advanced.Rmd:136
  repo's      advanced.Rmd:192
  structre    intermediate.Rmd:191
  Ts          advanced.Rmd:211
  webiste     advanced.Rmd:197
  ```

### help

- [ ] Again, the documentation for unexported function is confusing as functionality in the documentation doesn't work when you try and run it unless you know how to access internal functions.
- [ ] No examples found for `docker_build`,  `docker_push`, `git_upload`, `module_identities`, `module_skeleton`, `module_travis`, `module_upload`,
- [ ] how do `pkgdetails_get` and `description_get` differ?

- [ ] In `docker_build` help, what does value logical mean. ie what would it mean if the function returned `TRUE`?

### Test coverage.

- [ ] Great coverage overall. Can I just ask, is there a reason for the moderate coverage in `docker.R`?
  ```
  outsider.devtools Coverage: 90.36%
  R/docker.R: 56.25%
  ```

- [ ] Currently if you rerun `module_skeleton` using the same `flpth` argument, it ovewrites all files without warning. You might want to look at  checking for this and asking users to confirm overwrite (eg have a look at function `usethis::write_over()`).
- [ ] It feels like the `DESCRIPTION` file should have an `Author` field for a named person to formally take responsibility for the module. Using this would also allow users to create citations for their modules via `citation()`.
- [ ] Why is there a latest folder for a Dockerfiles? It seems that maybe the tag feature for versioning images is being mixed up wih versioning dockerfiles? As dockerfiles are just text files, they tend to generally be managed through version control.
- [ ] Why is there non-exported copies of `docker_build` and other `docker_cmd` in both `outsider.base`.
- [ ] It appears that using `module_build` with `build_image = FALSE` installs the module as a package in your local R library. In which situation would this be useful? It also means you don't need to build an image to use the functions in a module. You could be more explicit about what module building with and without building the image achieves. Perhaps you could point to and elaborate on these points in the FAQ section?

Some of these suggestions are minor and just personal opinions so feel free to ignore if you don't agree.

- [ ] Installation instructions in README should be changed as there is no need to manually install packages as you now have specified remotes.
- [ ] (Super minor!) In the quick guide block of code, it would be more legible with blank lines in between lines of code, eg:
  ```r
  library(outsider.devtools)

  # construct a skeleton file structure for the module
  module_path <- module_skeleton(program_name = 'echo', flpth = getwd())

  # check the file structure
  module_check(flpth = module_path)

  # look-up key identifying names: R package name, Docker image name
  module_identities(flpth = module_path)

  # build the R package and Docker image
  module_build(flpth = module_path, tag = 'latest')

  # test the module
  module_test(flpth = module_path)
  ```
- [ ] Installation instructions in the [Install](https://antonellilab.github.io/outsider.devtools/#install) section should be amended to reflect the automated installation of `outsider` and `outsider.base`.

- [ ]
> These code snippets are called Docker images.

  Not sure that's an accurate description of what a docker image is. My understanding is is that a docker image is an executable with an environment configured within which to run it. Overall and throughout documentation, I suggest you try to be precise with the terms you use in the docs and be consistent throughout. Docker is already somewhat confusing so misconceptions and inaccuracies in `outsider` docs could make it even more confusing. If unsure, have a look at the docker docs themselves.

##### Basic

- [ ] Very minor point but visually, a list structure would be a lot more effective:

   > This process comes in a series of steps:
   > 1. generate core files and folders
   > 1. create Docker image
   > 1. document and build R package
   > 1. try the module
   > 1. upload to GitHub and Docker Hub

- [ ]

  > The above code will create an outsider module with the name om..echo

  I feel it would be useful to point out here that the module is created at `flpth`.

- [ ] I know it feels like repetition but nevertheless, I feel it would be useful to show the resulting structure of the module from running `module_skeleton()` using output of command line `tree <temp-folder-path>/om..echo` :
  /var/folders/8p/87cqdx2s34vfvcgh04l6z72w0000gn/T//RtmptA6YLp/om..echo
  ├── DESCRIPTION
  ├── R
  │   ├── functions.R
  │   └── import.R
  ├── README.md
  ├── examples
  │   └── example.R
  └── inst
      ├── dockerfiles
      │   └── latest
      │       └── Dockerfile
      └── om.yml

  5 directories, 7 files
  ```

- [ ] The import and use of `outsider.base` functions in documentation or workflows is confusing because there is no function documentation in the `outsider.devtools` `pkgdown` reference section yet the package from which eg `outsider_init()` is imported from is not obvious to the user. I think it might be best to use the `reexports` approach (eg see how it [looks in the docs](https://devtools.r-lib.org/reference/reexports.html) and handled through roxygen in the [`R scripts`](https://github.com/r-lib/devtools/blob/master/R/inst.R) in devtools)

- [ ]
  >  Additionally, the developer would then not need to document all the arguments of the external program. For many external programs there may be hundreds of arguments, all of which are likely to be already documented;

  This is a great approach if documentation exists for the external module. But it might be good to encourage folks to check this and if there is none as part of the program, to atleast linked to appropriate documentation in the module docs.
- [ ]

  > For more information, look up “object documentation in R”.

  Best to provide the actual link to more information: http://r-pkgs.had.co.nz/man.html

- [ ]
  >  The WORKDIR sets the starting directory when a command is passed to the Docker image.

  This sentence feels confusing. It sets the **working directory**.  As I commented above, try to be precise and consistent with tricky terminology and if unsure, have a look at the docker docs themselves. eg here's the intro to `WORKDIR` instruction docs.

  > The WORKDIR instruction sets the working directory for any RUN, CMD, ENTRYPOINT, COPY and ADD instructions that follow it in the Dockerfile

- [ ]

  > Using our Dockerfile that has been preloaded with the installed R package

  Nothing is really _preloaded_ in a dockerfile. It is just a text file. The computational environment is configured during the buiding of the docker image. I personally would rewrite this paragraph instead describing step by step what happens when the dockerfile builds.

#### Advanced vignette:

- [ ]
 > To demonstrate the program’s complexity, here is the **command’s signature**

 That's rather unusual terminology. Do you mean **command's syntax**? Again I suggest you keep to conventional terminology.
