### 2020-01-13 #282 @DomBennett
Thanks very much @annakrystalli !

I've fixed a few of your comments already will get to work on the others soon.

- [ ] gp checks
- [ ] spell checks
- [ ] drop private function help topics
- [ ] create examples for: docker_build, docker_push, git_upload, module_identities, module_skeleton, module_travis, module_upload
- [ ] clarify distinction between `pkgdetails_get` and `description_get`
- [ ] clarify meaning of returned FALSE for `docker_` functions
- [ ] explain why docker functions test coverage is so low
- [ ] prevent overwriting pre-exisiting files with `module_skeleton`
- [ ] add author field to DESCRIPTION
- [ ] explain why there is a "latest" Dockerfile
- [ ] explain why there are non-exported copies of `docker_build` and other `docker_cmd` in both `outsider.base`
- [ ] explain the purpose of `build_image=FALSE` in `module_build`
- [ ] update install instructions in README
- [ ] Rephrase "These code snippets are called Docker images." from Basic vignette
- [ ] Create a list for steps in Basic vignette
- [ ] "The above code will create an outsider module with the name om..echo" indicate module exists at `flpth`
- [ ] Create tree diagram for echo module file structure
- [ ] Look into `outsider.base` import/exports. Attempt to clarify and normalise by mimicking well-known package structures
- [ ] Encourage better documentation of module arguments
- [ ] Add link: http://r-pkgs.had.co.nz/man.html
- [ ] Make more precise Docker terminology (in at least three locations)
- [ ] Command signature to syntax
