### 2020-01-15 #282 @DomBennett
Thanks @nuest !

@annakrystalli I've gone through my checklist and made all the requested changes (I hope!). Notable updates are:

* Now using reexports as suggested, see https://antonellilab.github.io/outsider.devtools/reference/reexports.html
* Added a lot more Docker context and have become a lot more careful about wording, e.g. see https://antonellilab.github.io/outsider.devtools/articles/basic.html#docker-files-and-folders
* No more accidental overwriting with `module_skeleton`, using `usethis::write_over`
* Skeleton allows developers to add their name and email
* Missing examples have been added
* Description for why "latest" tag has been added to "basic" article

### Clarifications

* `docker_build` and `docker_cmd` exisiting in both the dev and base packages. They are both functionally the same, but in the dev version they print to console the bash commands to make it more transparent what is going on.
* There are no strong reasons to set `build_image=FALSE`. I do this in the example to separate the R building from the image building so that I don't need to explain both processes in the same section. One reason to avoid building an image is if you are updating the package code but haven't changed the Dockerfile. Would that be useful info to add to the FAQs?

### Other

* There is now an online docs for the `outsider.base`, https://antonellilab.github.io/outsider.base/reference/index.html. I was hoping the devtools ["reexports" page](https://antonellilab.github.io/outsider.devtools/reference/reexports.html) would automatically link to this documentation rather than the rddr.io documentation but I'm not sure how `pkgdown` links.

---

Let me know if you need me to do anything else!
