### 2020-01-16 #282 @DomBennett
Huzzah!

Thanks so much @annakrystalli!

Also big thanks to @nuest and @hrbrmstr!

Is it OK to add you all as "rev" in the R package DESCRIPTION?
