### 2020-01-16 #282 @annakrystalli
Hey @DomBennett, thanks for transferring the repos over, I've returned full admin rights back to you.

Just double-checking with the editors team whether all will live in rOpenSci. In the meantime, could you briefly explain what the `outsider-testsuite` repo is/contains? Is it user facing?
