### 2020-01-16 #282 @DomBennett
Thanks!

`outsider-testsuite` is not user facing. In the review one of the suggestions was to have a demonstrated pipeline of how multiple modules could be stringed together to test the overall `outsider` "universe".

Currently, there is only 1 functioning test suite for a phylogenetic pipeline. I do invite people to contribute their own suites (but now i'm now sure where I wrote that..... 😕)

(It seemed a bit much to have so many tests in one of the packages.)
