### 2020-01-16 ropensci/outsider#6 @DomBennett
Docker running, all passing.
