### 2020-01-16 #282 @annakrystalli
OK! I've heard back and all is good. All 4 repos are fine in rOpenSci and as part of the `outsider` team. Just make sure (which currently you have) that the peer-reviewed badge is only on the two packages.

I'm pretty much ready to close the issue once you confirm all the approval checklist items have been tackled. 😃
