### 2020-01-23 #282 @DomBennett
Now published: https://joss.theoj.org/papers/10.21105/joss.02038#

I've removed @nuest and @hrbrmstr as "rev" in the DESCRIPTION file for now just to get things moving along. Happy to add you at a later date, of course!

Thanks again to you all for making this happen!
