### 2020-01-25 #282 @annakrystalli
Great news on the JOSS publication. As for reviewer acknowledgement, that seems fair enough. Reviewers just ping in this issue if you would like to be officially acknowledged.

Going to close the issue and officially conclude the review. Great work again!
